func stat_avg_count is work.
