+++ src/addit_func.py
@@ -1,4 +1,5 @@
 import os
+from datetime import datetime
 
 
 def scan_dirs(path: str, recurse: bool = False) -> dict:
@@ -84,3 +85,52 @@ def sort_dict_list(product_list: list, cat: str | None) -> list:
 
     sorted_product = sorted(out_prod_list, key=lambda x: x["price"], reverse=True)
     return sorted_product
+
+
+def sorting_list(raw_list: list) -> list:
+    """функция принимает список,подсчитывае сумму в заказе,
+    сортирует список по дате и возвращает в виде списка
+    значений [{'date': '2024-01', 'sum': 41.5},...]"""
+    temp_list = []
+    for item in raw_list:
+        temp_dic = {}
+        date2 = (datetime.strptime(item["date"], "%d.%m.%Y")).strftime("%Y-%m")
+        sum1 = 0
+        for item2 in item.get("items"):
+            sum1 = sum1 + item2["price"] * item2["quantity"]
+        temp_dic["date"] = date2
+        temp_dic["sum"] = sum1
+        temp_list.append(temp_dic)
+    return sorted(temp_list, key=lambda x: x["date"])
+
+
+def summ_price(products: list, category: str) -> dict:
+    """функция суммирует значения в указанной категории,возвращает среднее значение,
+    и количество вхождений."""
+    avg2 = {}
+    summ = 0
+    count = 0
+    for product in products:
+        if product["date"] == category:
+            summ += product["sum"]
+            count += 1
+    avg = round(summ / count, 2)
+    avg2["average_order_value"] = avg
+    avg2["order_count"] = count
+
+    return avg2
+
+
+def stat_avg_count(stat_list: list) -> dict:
+    """функция принимает список словарей,возвращает список словарь
+    вида {'2024-01': {'average_order_value': 222.32, 'order_count': 4},
+    где указаны средняя сумма заказов,их количество. Ключом выступает месяц"""
+    temp_list1 = sorting_list(stat_list)
+    avg1 = {}
+    data_set1 = []
+    for item in temp_list1:
+        if not item["date"] in data_set1:
+            data_set1.append(item["date"])
+    for dates in data_set1:
+        avg1[dates] = summ_price(temp_list1, dates)
+    return avg1
